Restructure file imports
