+++ src/components/AboutImportantHeader.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import { withStyles } from "@material-ui/core";
 
-import Gill from "../assets/aboutPage/gill.png";
+import Gill from "../assets/gill.png";
 
 const styles = {
   parentContainer: {
