+++ src/components/AllStoriesButton.js
@@ -2,7 +2,7 @@ import React from "react";
 import { withStyles } from "@material-ui/core";
 import PropTypes from "prop-types";
 
-import arrow from "../assets/shared/arrowBlueRight.png";
+import arrow from "../assets/arrowBlueRight.png";
 
 import TextArrowButton from "./TextArrowButton";
