+++ src/components/MonthStory/MonthStoryText.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import { withStyles, Typography } from "@material-ui/core";
 
-import ArrowRight from "../../assets/shared/arrowWhite.png";
+import ArrowRight from "../../assets/arrowWhite.png";
 
 const styles = {
   monthStoryParent: {
