+++ src/components/More.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import { Grid, Typography, withStyles } from "@material-ui/core";
 
-import arrow from "../assets/shared/arrowBlueRight.png";
+import arrow from "../assets/arrowBlueRight.png";
 
 import TextArrowButton from "./TextArrowButton";
